requirements.txt
fastapi
uvicorn
langgraph
langchain
langchain-anthropic
langchain-community
chromadb
python-dotenv
pydantic
sentence-transformers
mcp
httpx
